Rename `AddBattlerTagIfBoostedAttr`
