+++ src/data/init/init-moves.ts
@@ -32,7 +32,6 @@ import { AbilityCopyAttr } from "#moves/ability-copy-attr";
 import { AbilityGiveAttr } from "#moves/ability-give-attr";
 import { AcupressureStatStageChangeAttr } from "#moves/acupressure-stat-stage-change-attr";
 import { AddBattlerTagHeaderAttr } from "#moves/add-battler-tag-header-attr";
-import { ConfuseIfBoostedAttr } from "#moves/add-battler-tag-if-boosted-attr";
 import { AddEntryHazardTagAttr } from "#moves/add-entry-hazard-tag-attr";
 import { AddPledgeEffectAttr } from "#moves/add-pledge-effect-attr";
 import { AddSubstituteAttr } from "#moves/add-substitute-attr";
@@ -74,6 +73,7 @@ import { CombinedPledgeTypeAttr } from "#moves/combined-pledge-type-attr";
 import { CompareWeightPowerAttr } from "#moves/compare-weight-power-attr";
 import { ConditionalProtectAttr } from "#moves/conditional-protect-attr";
 import { ConfuseAttr } from "#moves/confuse-attr";
+import { ConfuseIfBoostedAttr } from "#moves/confuse-if-boosted-attr";
 import { ConsecutiveUseDoublePowerAttr } from "#moves/consecutive-use-double-power-attr";
 import { ConsecutiveUseMultiBasePowerAttr } from "#moves/consecutive-use-multi-base-power-attr";
 import { CopyBiomeTypeAttr } from "#moves/copy-biome-type-attr";
