+++ src/data/moves/move-attrs/confuse-if-boosted-attr.ts
